basic user and group serializer and view
