+++ community/rest_api/serializers/__init__.py
@@ -1,0 +1,1 @@
+from . import user, group
