+++ community/rest_api/serializers/group.py
@@ -1,0 +1,7 @@
+from django.contrib.auth.models import Group
+from rest_framework import serializers
+
+class GroupSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Group
+        fields = ('url', 'name')
