+++ community/rest_api/serializers/user.py
@@ -1,0 +1,16 @@
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+
+    def create(self, validated_data):
+        return User.objects.create(**validated_data)
+
+    def update(self, instance, validated_data):
+        instance.save()
+        return instance
+
+    class Meta:
+        model = User
+        fields = ('url', 'username', 'email', 'groups')
