--- community/rest_api/views.py
@@ -1,19 +1,0 @@
-from django.contrib.auth.models import User, Group
-from rest_framework import viewsets
-from community.rest_api.serializers import UserSerializer, GroupSerializer
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    """
-    API endpoint that allows users to be viewed or edited.
-    """
-    queryset = User.objects.all().order_by('-date_joined')
-    serializer_class = UserSerializer
-
-
-class GroupViewSet(viewsets.ModelViewSet):
-    """
-    API endpoint that allows groups to be viewed or edited.
-    """
-    queryset = Group.objects.all().order_by('-name')
-    serializer_class = GroupSerializer
